+++ include/Core/World/ChunkMesher.hpp
@@ -10,13 +10,13 @@ namespace kc {
     public:
         static void generateSubChunk(
             SubChunk& subChunk,
-            const std::vector<uint8_t>& blockIds,
+            const std::vector<uint64_t>& blockIds,
             int N, int height,
             int yBase,
-            const std::vector<uint8_t>* edgePosX = nullptr,
-            const std::vector<uint8_t>* edgeNegX = nullptr,
-            const std::vector<uint8_t>* edgePosZ = nullptr,
-            const std::vector<uint8_t>* edgeNegZ = nullptr);
+            const std::vector<uint64_t>* edgePosX = nullptr,
+            const std::vector<uint64_t>* edgeNegX = nullptr,
+            const std::vector<uint64_t>* edgePosZ = nullptr,
+            const std::vector<uint64_t>* edgeNegZ = nullptr);
 
         // Emits ONLY the faces of subChunk pointing in the given gate direction
         // (2=PosZ, 3=NegZ, 4=PosX, 5=NegX). Used for border re-culling: when a
@@ -25,14 +25,14 @@ namespace kc {
         // of rebuilding all 25 sub-chunks. Appends to subChunk.vertices/indices.
         static void emitGateFaces(
             SubChunk& subChunk,
-            const std::vector<uint8_t>& blockIds,
+            const std::vector<uint64_t>& blockIds,
             int N, int height,
             int yBase,
             int gate,
-            const std::vector<uint8_t>* edgePosX = nullptr,
-            const std::vector<uint8_t>* edgeNegX = nullptr,
-            const std::vector<uint8_t>* edgePosZ = nullptr,
-            const std::vector<uint8_t>* edgeNegZ = nullptr);
+            const std::vector<uint64_t>* edgePosX = nullptr,
+            const std::vector<uint64_t>* edgeNegX = nullptr,
+            const std::vector<uint64_t>* edgePosZ = nullptr,
+            const std::vector<uint64_t>* edgeNegZ = nullptr);
     };
 
 } // namespace kc
