+++ include/Core/World/Physics/CollisionSystem.hpp
@@ -12,8 +12,8 @@ namespace kc {
     class CollisionSystem {
     public:
         static bool isSolidBlock(const World& world, int x, int y, int z);
-        static const AABB& getBlockCollisionBox(uint8_t blockId);
-        static AABB blockAABBAt(uint8_t blockId, int x, int y, int z);
+        static const AABB& getBlockCollisionBox(uint64_t blockId);
+        static AABB blockAABBAt(uint64_t blockId, int x, int y, int z);
         static AABB blockAABBAt(const Block& block, int x, int y, int z);
         static std::optional<AABB> getWorldBlockAABB(const World& world, int x, int y, int z);
         static bool aabbCollides(const World& world, const AABB& box);
